move test from unit to integration to save time
